+++ include/ivi-logging-stl.h
@@ -16,12 +16,12 @@ namespace logging {
 template<typename ElementType, class LogDataType = logging::LogData, typename =
 		 typename std::enable_if<std::is_base_of<logging::LogData, LogDataType>::value>::type>
 LogDataType& operator<<(LogDataType& log, const std::vector<ElementType>& v) {
-	log << " [ ";
+	log << "[ ";
 	for (auto& element : v) {
 		log << element;
 		log << ", ";
 	}
-	log << " ] ";
+	log << "]";
 	return log;
 }